Support decimal quantities and prices for B2B/B2C sales.
Allow fractional sale quantities (e.g. 2.5 meters) with DECIMAL(12,3) columns, decimal stock deduction, price rounding to two decimals, and serial numbers only when quantity is a whole number.

+++ models/InventoryTransaction.ts
@@ -44,7 +44,7 @@ InventoryTransaction.init(
       allowNull: false
     },
     quantity: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(12, 3),
       allowNull: false
     },
     reference: {
